Add Cobra command-line parser
