+++ cmd/server/server.go
@@ -1,0 +1,7 @@
+package main
+
+import "github.com/project-todo/todo-list-api/cmd/server/cmd"
+
+func main() {
+	cmd.Execute()
+}
