+++ go.mod
@@ -2,4 +2,7 @@ module github.com/project-todo/todo-list-api
 
 go 1.12
 
-require github.com/gorilla/mux v1.7.3
+require (
+	github.com/gorilla/mux v1.7.3
+	github.com/spf13/cobra v0.0.5
+)
